v3-fixed

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jesssanc <jesssanc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/09 13:33:25 by jesssanc          #+#    #+#             */
-/*   Updated: 2024/12/12 09:33:35 by jesssanc         ###   ########.fr       */
+/*   Updated: 2024/12/12 10:02:52 by jesssanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	ft_printfchar(char c);
 int	ft_printstr(char *s);
 int	ft_printnbr(int n);
 int	ft_printunnbr(unsigned int n);
-int	ft_printhex_low(unsigned long long num);
-int	ft_printhex_up(unsigned long long num);
+int	ft_printhex_low(unsigned long num);
+int	ft_printhex_up(unsigned long num);
 
 #endif
